Sofi

+++ controlador/ControlPrestamo.php
@@ -14,25 +14,55 @@
         public function irDevolver(){            
             require_once("vista/devolver.php");
         }
-        public function agregarPrestamo($id, $fprestamo,  $fdevolucion, $sesion){
-            $usuario = $sesion->getUsuario();
-           $resultado = $this->mprestamo->agregarPrestamo($id, $fprestamo,  $fdevolucion, $usuario);
-           if($resultado){
-            $sesion->setMensaje("Prestamo agregado");
-          }else{
-            $sesion->setMensaje("error agregando el prestamo");
+
+        public function agregarPrestamo($id, $fprestamo, $fdevolucion, $sesion) {
+          $usuario = $sesion->getUsuario();
+      
+          // Verificar si existe un libro con el ID proporcionado
+          if ($this->mprestamo->existeLibro($id)) {
+              // Verificar si el libro está disponible para préstamo
+              if ($this->mprestamo->libroDisponible($id)) {
+                  $resultado = $this->mprestamo->agregarPrestamo($id, $fprestamo, $fdevolucion, $usuario);
+                  if ($resultado) {
+                      $sesion->setMensaje("Préstamo agregado");
+                  } else {
+                      if (strtotime($fprestamo) > strtotime($fdevolucion)) {
+                          $sesion->setMensaje("La fecha de préstamo debe ser anterior a la fecha de devolución");
+                      } else {
+                          $sesion->setMensaje("Error agregando el préstamo");
+                      }
+                  }
+              } else {
+                  $sesion->setMensaje("El libro no está disponible para préstamo");
+              }
+          } else {
+              $sesion->setMensaje("No existe un libro con ese ID");
           }
-            require_once("vista/alta_prestamo.php");
-        }
+      
+          require_once("vista/alta_prestamo.php");
+      }
+      
+      
+      
+    
+      
 
         public function eliminarPrestamo($id, $sesion){
             $usuario = $sesion->getUsuario();
-           $resultado = $this->mprestamo->eliminarPrestamo($id, $sesion);
-           if($resultado){
-            $sesion->setMensaje("Prestamo eliminado");
-          }else{
-            $sesion->setMensaje("error eliminando el prestamo");
-          }
+            if ($this->mprestamo->existeLibro($id)) {
+              if($this->mprestamo->exists($id)){
+              $resultado = $this->mprestamo->eliminarPrestamo($id, $sesion);
+              if($resultado){
+                $sesion->setMensaje("Prestamo devuelto");
+              }else{
+                $sesion->setMensaje("Error devolviendo el prestamo");
+              }
+            }else{
+              $sesion->setMensaje("Error en devolucion: El libro ya se encuentra disponible"); 
+            }
+            }else{
+              $sesion->setMensaje("No existe un libro con ese ID");
+            }
 
           require_once("vista/devolver.php");
         }
